weave-backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests
pydantic
python-multipart
